requirements.txt
# Core
websockets>=12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0

# Task 1 — Tabular ML
scikit-learn>=1.4.0
lightgbm>=4.3.0

# Task 2 — NLP
transformers>=4.40.0
torch>=2.2.0
sentencepiece>=0.2.0

# Task 3 — Computer Vision
torchvision>=0.17.0
Pillow>=10.0.0

# Utilities
joblib>=1.3.0
scipy>=1.12.0
